test: rewrite keep alive tests
