+++ crates/curp/src/client/keep_alive.rs
@@ -135,3 +135,216 @@ impl KeepAlive {
             .await
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::collections::HashMap;
+
+    use super::*;
+
+    use futures::{future::BoxFuture, Stream};
+    use tonic::Status;
+    use tracing_test::traced_test;
+
+    use crate::rpc::{
+        connect::{ConnectApi, MockConnectApi},
+        CurpError, FetchClusterRequest, FetchClusterResponse, FetchReadStateRequest,
+        FetchReadStateResponse, Member, MoveLeaderRequest, MoveLeaderResponse, OpResponse,
+        ProposeConfChangeRequest, ProposeConfChangeResponse, ProposeId, ProposeRequest,
+        ProposeResponse, PublishRequest, PublishResponse, ReadIndexResponse, RecordRequest,
+        RecordResponse, ResponseOp, ShutdownRequest, ShutdownResponse, SyncedResponse,
+    };
+
+    struct MockedStreamConnectApi {
+        id: u64,
+        lease_keep_alive_handle:
+            Box<dyn Fn(u64) -> BoxFuture<'static, Result<u64, CurpError>> + Send + Sync + 'static>,
+    }
+
+    #[async_trait::async_trait]
+    impl ConnectApi for MockedStreamConnectApi {
+        /// Get server id
+        fn id(&self) -> u64 {
+            self.id
+        }
+
+        /// Update server addresses, the new addresses will override the old ones
+        async fn update_addrs(&self, _addrs: Vec<String>) -> Result<(), tonic::transport::Error> {
+            Ok(())
+        }
+
+        /// Send `ProposeRequest`
+        async fn propose_stream(
+            &self,
+            _request: ProposeRequest,
+            _token: Option<String>,
+            _timeout: Duration,
+        ) -> Result<
+            tonic::Response<Box<dyn Stream<Item = Result<OpResponse, Status>> + Send>>,
+            CurpError,
+        > {
+            unreachable!("please use MockedConnectApi")
+        }
+
+        /// Send `RecordRequest`
+        async fn record(
+            &self,
+            _request: RecordRequest,
+            _timeout: Duration,
+        ) -> Result<tonic::Response<RecordResponse>, CurpError> {
+            unreachable!("please use MockedConnectApi")
+        }
+
+        /// Send `ReadIndexRequest`
+        async fn read_index(
+            &self,
+            _timeout: Duration,
+        ) -> Result<tonic::Response<ReadIndexResponse>, CurpError> {
+            unreachable!("please use MockedConnectApi")
+        }
+
+        /// Send `ProposeConfChange`
+        async fn propose_conf_change(
+            &self,
+            _request: ProposeConfChangeRequest,
+            _timeout: Duration,
+        ) -> Result<tonic::Response<ProposeConfChangeResponse>, CurpError> {
+            unreachable!("please use MockedConnectApi")
+        }
+
+        /// Send `PublishRequest`
+        async fn publish(
+            &self,
+            _request: PublishRequest,
+            _timeout: Duration,
+        ) -> Result<tonic::Response<PublishResponse>, CurpError> {
+            unreachable!("please use MockedConnectApi")
+        }
+
+        /// Send `ShutdownRequest`
+        async fn shutdown(
+            &self,
+            _request: ShutdownRequest,
+            _timeout: Duration,
+        ) -> Result<tonic::Response<ShutdownResponse>, CurpError> {
+            unreachable!("please use MockedConnectApi")
+        }
+
+        /// Send `FetchClusterRequest`
+        async fn fetch_cluster(
+            &self,
+            _request: FetchClusterRequest,
+            _timeout: Duration,
+        ) -> Result<tonic::Response<FetchClusterResponse>, CurpError> {
+            let members = (0..5)
+                .into_iter()
+                .map(|id| Member::new(id, format!("{id}"), vec![], vec![], false))
+                .collect();
+            let resp = FetchClusterResponse::new(Some(0), 1, 1, members, 1);
+            Ok(tonic::Response::new(resp))
+        }
+
+        /// Send `FetchReadStateRequest`
+        async fn fetch_read_state(
+            &self,
+            _request: FetchReadStateRequest,
+            _timeout: Duration,
+        ) -> Result<tonic::Response<FetchReadStateResponse>, CurpError> {
+            unreachable!("please use MockedConnectApi")
+        }
+
+        /// Send `MoveLeaderRequest`
+        async fn move_leader(
+            &self,
+            _request: MoveLeaderRequest,
+            _timeout: Duration,
+        ) -> Result<tonic::Response<MoveLeaderResponse>, CurpError> {
+            unreachable!("please use MockedConnectApi")
+        }
+
+        /// Keep send lease keep alive to server and mutate the client id
+        async fn lease_keep_alive(
+            &self,
+            client_id: u64,
+            _interval: Duration,
+        ) -> Result<u64, CurpError> {
+            (self.lease_keep_alive_handle)(client_id).await
+        }
+    }
+
+    /// Create mocked stream connects
+    ///
+    /// The leader is S0
+    #[allow(trivial_casts)] // cannot be inferred
+    fn init_mocked_stream_connects(
+        size: usize,
+        leader_idx: usize,
+        leader_term: u64,
+        keep_alive_handle: impl Fn(u64) -> BoxFuture<'static, Result<u64, CurpError>>
+            + Send
+            + Sync
+            + 'static,
+    ) -> HashMap<u64, Arc<dyn ConnectApi>> {
+        let mut keep_alive_handle = Some(keep_alive_handle);
+        let redirect_handle = move |_id| {
+            Box::pin(async move { Err(CurpError::redirect(Some(leader_idx as u64), leader_term)) })
+                as BoxFuture<'static, Result<u64, CurpError>>
+        };
+        (0..size)
+            .map(|id| MockedStreamConnectApi {
+                id: id as u64,
+                lease_keep_alive_handle: if id == leader_idx {
+                    Box::new(keep_alive_handle.take().unwrap())
+                } else {
+                    Box::new(redirect_handle)
+                },
+            })
+            .enumerate()
+            .map(|(id, api)| (id as u64, Arc::new(api) as Arc<dyn ConnectApi>))
+            .collect()
+    }
+
+    /// Create stream client for test
+    fn init_stream_client(
+        connects: HashMap<u64, Arc<dyn ConnectApi>>,
+        leader: u64,
+        term: u64,
+        cluster_version: u64,
+    ) -> KeepAliveHandle {
+        let state = ClusterState::Ready(ClusterStateReady::new(
+            leader,
+            term,
+            cluster_version,
+            connects.clone(),
+        ));
+        let fetch = Fetch::new(Duration::from_secs(0), move |_| connects.clone());
+        let state_shared = ClusterStateShared::new_test(state, fetch);
+
+        let keep_alive = KeepAlive::new(Duration::from_secs(1));
+        keep_alive.spawn_keep_alive(Arc::new(state_shared))
+    }
+
+    #[traced_test]
+    #[tokio::test]
+    async fn test_stream_client_keep_alive_works() {
+        let connects =
+            init_mocked_stream_connects(5, 0, 1, move |client_id| Box::pin(async move { Ok(10) }));
+        let mut keep_alive = init_stream_client(connects, 0, 1, 1);
+        tokio::time::timeout(Duration::from_millis(100), &mut keep_alive.handle)
+            .await
+            .unwrap_err();
+        assert_eq!(keep_alive.wait_id_update(0).await, 10);
+    }
+
+    #[traced_test]
+    #[tokio::test]
+    async fn test_stream_client_keep_alive_on_redirect() {
+        let connects =
+            init_mocked_stream_connects(5, 0, 2, move |client_id| Box::pin(async move { Ok(10) }));
+        let mut keep_alive = init_stream_client(connects, 1, 1, 1);
+        tokio::time::timeout(Duration::from_millis(100), &mut keep_alive.handle)
+            .await
+            .unwrap_err();
+        assert_eq!(keep_alive.wait_id_update(0).await, 10);
+    }
+}
